feat: promote backstage-cli-alpha to the main entrypoint
