+++ packages/cli/package.json
@@ -20,8 +20,7 @@
   "license": "Apache-2.0",
   "main": "dist/index.cjs.js",
   "bin": {
-    "backstage-cli": "bin/backstage-cli",
-    "backstage-cli-alpha": "bin/backstage-cli-alpha"
+    "backstage-cli": "bin/backstage-cli"
   },
   "files": [
     "asset-types",
