--- packages/cli/src/modules/create-github-app/alpha.ts
@@ -1,38 +1,0 @@
-/*
- * Copyright 2025 The Backstage Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import { createCliPlugin } from '../../wiring/factory';
-import { Command } from 'commander';
-import { lazy } from '../../lib/lazy';
-
-export default createCliPlugin({
-  pluginId: 'new',
-  init: async reg => {
-    reg.addCommand({
-      path: ['create-github-app'],
-      description: 'Create new GitHub App in your organization.',
-      execute: async ({ args }) => {
-        const command = new Command();
-        const defaultCommand = command
-          .argument('<github-org>')
-          .action(
-            lazy(() => import('./commands/create-github-app'), 'default'),
-          );
-
-        await defaultCommand.parseAsync(args, { from: 'user' });
-      },
-    });
-  },
-});
