+++ packages/cli/src/modules/info/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Backstage Authors
+ * Copyright 2024 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,13 +13,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+import yargs from 'yargs';
+import { createCliPlugin } from '../../wiring/factory';
 
-import { lazy } from '../../lib/lazy';
-import { Command } from 'commander';
-
-export function registerCommands(program: Command) {
-  program
-    .command('info')
-    .description('Show helpful information for debugging and reporting bugs')
-    .action(lazy(() => import('./commands/info'), 'default'));
-}
+export default createCliPlugin({
+  pluginId: 'info',
+  init: async reg => {
+    reg.addCommand({
+      path: ['info'],
+      description: 'Show helpful information for debugging and reporting bugs',
+      execute: async ({ args }) => {
+        yargs().parse(args);
+        const { default: command } =
+          require('./commands/info') as typeof import('./commands/info');
+        await command();
+      },
+    });
+  },
+});
