+++ packages/cli/src/modules/maintenance/index.ts
@@ -13,49 +13,98 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 import { Command } from 'commander';
+import { createCliPlugin } from '../../wiring/factory';
 import { lazy } from '../../lib/lazy';
 
-export function registerPackageCommands(command: Command) {
-  command
-    .command('clean')
-    .description('Delete cache directories')
-    .action(lazy(() => import('./commands/package/clean'), 'default'));
+export default createCliPlugin({
+  pluginId: 'maintenance',
+  init: async reg => {
+    reg.addCommand({
+      path: ['package', 'clean'],
+      description: 'Delete cache directories',
+      execute: async ({ args }) => {
+        const command = new Command();
+        const defaultCommand = command.action(
+          lazy(() => import('./commands/package/clean'), 'default'),
+        );
 
-  command
-    .command('prepack')
-    .description('Prepares a package for packaging before publishing')
-    .action(lazy(() => import('./commands/package/pack'), 'pre'));
+        await defaultCommand.parseAsync(args, { from: 'user' });
+      },
+    });
 
-  command
-    .command('postpack')
-    .description('Restores the changes made by the prepack command')
-    .action(lazy(() => import('./commands/package/pack'), 'post'));
-}
+    reg.addCommand({
+      path: ['package', 'prepack'],
+      description: 'Prepares a package for packaging before publishing',
+      execute: async ({ args }) => {
+        const command = new Command();
+        const defaultCommand = command.action(
+          lazy(() => import('./commands/package/pack'), 'pre'),
+        );
 
-export function registerRepoCommands(command: Command) {
-  command
-    .command('fix')
-    .description('Automatically fix packages in the project')
-    .option(
-      '--publish',
-      'Enable additional fixes that only apply when publishing packages',
-    )
-    .option(
-      '--check',
-      'Fail if any packages would have been changed by the command',
-    )
-    .action(lazy(() => import('./commands/repo/fix'), 'command'));
+        await defaultCommand.parseAsync(args, { from: 'user' });
+      },
+    });
 
-  command
-    .command('clean')
-    .description('Delete cache and output directories')
-    .action(lazy(() => import('./commands/repo/clean'), 'command'));
+    reg.addCommand({
+      path: ['package', 'postpack'],
+      description: 'Restores the changes made by the prepack command',
+      execute: async ({ args }) => {
+        const command = new Command();
+        const defaultCommand = command.action(
+          lazy(() => import('./commands/package/pack'), 'post'),
+        );
 
-  command
-    .command('list-deprecations')
-    .description('List deprecations')
-    .option('--json', 'Output as JSON')
-    .action(lazy(() => import('./commands/repo/list-deprecations'), 'command'));
-}
+        await defaultCommand.parseAsync(args, { from: 'user' });
+      },
+    });
+
+    reg.addCommand({
+      path: ['repo', 'fix'],
+      description: 'Automatically fix packages in the project',
+      execute: async ({ args }) => {
+        const command = new Command();
+        const defaultCommand = command
+          .option(
+            '--publish',
+            'Enable additional fixes that only apply when publishing packages',
+          )
+          .option(
+            '--check',
+            'Fail if any packages would have been changed by the command',
+          )
+          .action(lazy(() => import('./commands/repo/fix'), 'command'));
+
+        await defaultCommand.parseAsync(args, { from: 'user' });
+      },
+    });
+
+    reg.addCommand({
+      path: ['repo', 'clean'],
+      description: 'Delete cache and output directories',
+      execute: async ({ args }) => {
+        const command = new Command();
+        const defaultCommand = command.action(
+          lazy(() => import('./commands/repo/clean'), 'command'),
+        );
+
+        await defaultCommand.parseAsync(args, { from: 'user' });
+      },
+    });
+
+    reg.addCommand({
+      path: ['repo', 'list-deprecations'],
+      description: 'List deprecations',
+      execute: async ({ args }) => {
+        const command = new Command();
+        const defaultCommand = command
+          .option('--json', 'Output as JSON')
+          .action(
+            lazy(() => import('./commands/repo/list-deprecations'), 'command'),
+          );
+
+        await defaultCommand.parseAsync(args, { from: 'user' });
+      },
+    });
+  },
+});
