+++ packages/cli/src/modules/migrate/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 The Backstage Authors
+ * Copyright 2024 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,72 +13,123 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-import { lazy } from '../../lib/lazy';
+import { createCliPlugin } from '../../wiring/factory';
 import { Command } from 'commander';
+import { lazy } from '../../lib/lazy';
 
-export function registerCommands(program: Command) {
-  program
-    .command('versions:bump')
-    .option(
-      '--pattern <glob>',
-      'Override glob for matching packages to upgrade',
-    )
-    .option(
-      '--release <version|next|main>',
-      'Bump to a specific Backstage release line or version',
-      'main',
-    )
-    .option('--skip-install', 'Skips yarn install step')
-    .option('--skip-migrate', 'Skips migration of any moved packages')
-    .description('Bump Backstage packages to the latest versions')
-    .action(lazy(() => import('./commands/versions/bump'), 'default'));
+export default createCliPlugin({
+  pluginId: 'migrate',
+  init: async reg => {
+    reg.addCommand({
+      path: ['versions:migrate'],
+      description:
+        'Migrate any plugins that have been moved to the @backstage-community namespace automatically',
+      execute: async ({ args }) => {
+        const command = new Command();
+        const defaultCommand = command
+          .option(
+            '--pattern <glob>',
+            'Override glob for matching packages to upgrade',
+          )
+          .option(
+            '--skip-code-changes',
+            'Skip code changes and only update package.json files',
+          )
+          .action(lazy(() => import('./commands/versions/migrate'), 'default'));
 
-  program
-    .command('versions:migrate')
-    .option(
-      '--pattern <glob>',
-      'Override glob for matching packages to upgrade',
-    )
-    .option(
-      '--skip-code-changes',
-      'Skip code changes and only update package.json files',
-    )
-    .description(
-      'Migrate any plugins that have been moved to the @backstage-community namespace automatically',
-    )
-    .action(lazy(() => import('./commands/versions/migrate'), 'default'));
+        await defaultCommand.parseAsync(args, { from: 'user' });
+      },
+    });
 
-  const command = program
-    .command('migrate [command]')
-    .description('Migration utilities');
+    reg.addCommand({
+      path: ['versions:bump'],
+      description: 'Bump Backstage packages to the latest versions',
+      execute: async ({ args }) => {
+        const command = new Command();
 
-  command
-    .command('package-roles')
-    .description(`Add package role field to packages that don't have it`)
-    .action(lazy(() => import('./commands/packageRole'), 'default'));
+        const defaultCommand = command
+          .option(
+            '--pattern <glob>',
+            'Override glob for matching packages to upgrade',
+          )
+          .option(
+            '--release <version|next|main>',
+            'Bump to a specific Backstage release line or version',
+            'main',
+          )
+          .option('--skip-install', 'Skips yarn install step')
+          .option('--skip-migrate', 'Skips migration of any moved packages')
+          .action(lazy(() => import('./commands/versions/bump'), 'default'));
 
-  command
-    .command('package-scripts')
-    .description('Set package scripts according to each package role')
-    .action(lazy(() => import('./commands/packageScripts'), 'command'));
+        await defaultCommand.parseAsync(args, { from: 'user' });
+      },
+    });
 
-  command
-    .command('package-exports')
-    .description('Synchronize package subpath export definitions')
-    .action(lazy(() => import('./commands/packageExports'), 'command'));
+    reg.addCommand({
+      path: ['migrate', 'package-roles'],
+      description: `Add package role field to packages that don't have it`,
+      execute: async ({ args }) => {
+        const command = new Command();
+        const defaultCommand = command.action(
+          lazy(() => import('./commands/packageRole'), 'default'),
+        );
 
-  command
-    .command('package-lint-configs')
-    .description(
-      'Migrates all packages to use @backstage/cli/config/eslint-factory',
-    )
-    .action(lazy(() => import('./commands/packageLintConfigs'), 'command'));
+        await defaultCommand.parseAsync(args, { from: 'user' });
+      },
+    });
 
-  command
-    .command('react-router-deps')
-    .description(
-      'Migrates the react-router dependencies for all packages to be peer dependencies',
-    )
-    .action(lazy(() => import('./commands/reactRouterDeps'), 'command'));
-}
+    reg.addCommand({
+      path: ['migrate', 'package-scripts'],
+      description: 'Set package scripts according to each package role',
+      execute: async ({ args }) => {
+        const command = new Command();
+        const defaultCommand = command.action(
+          lazy(() => import('./commands/packageScripts'), 'command'),
+        );
+
+        await defaultCommand.parseAsync(args, { from: 'user' });
+      },
+    });
+
+    reg.addCommand({
+      path: ['migrate', 'package-exports'],
+      description: 'Synchronize package subpath export definitions',
+      execute: async ({ args }) => {
+        const command = new Command();
+        const defaultCommand = command.action(
+          lazy(() => import('./commands/packageExports'), 'command'),
+        );
+
+        await defaultCommand.parseAsync(args, { from: 'user' });
+      },
+    });
+
+    reg.addCommand({
+      path: ['migrate', 'package-lint-configs'],
+      description:
+        'Migrates all packages to use @backstage/cli/config/eslint-factory',
+      execute: async ({ args }) => {
+        const command = new Command();
+        const defaultCommand = command.action(
+          lazy(() => import('./commands/packageLintConfigs'), 'command'),
+        );
+
+        await defaultCommand.parseAsync(args, { from: 'user' });
+      },
+    });
+
+    reg.addCommand({
+      path: ['migrate', 'react-router-deps'],
+      description:
+        'Migrates the react-router dependencies for all packages to be peer dependencies',
+      execute: async ({ args }) => {
+        const command = new Command();
+        const defaultCommand = command.action(
+          lazy(() => import('./commands/reactRouterDeps'), 'command'),
+        );
+
+        await defaultCommand.parseAsync(args, { from: 'user' });
+      },
+    });
+  },
+});
