+++ packages/cli/src/modules/new/index.ts
@@ -13,53 +13,74 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+import { createCliPlugin } from '../../wiring/factory';
 import { Command } from 'commander';
 import { lazy } from '../../lib/lazy';
 import { removed } from '../../lib/removed';
 
-export function registerCommands(program: Command) {
-  program
-    .command('new')
-    .storeOptionsAsProperties(false)
-    .description(
-      'Open up an interactive guide to creating new things in your app',
-    )
-    .option(
-      '--select <name>',
-      'Select the thing you want to be creating upfront',
-    )
-    .option(
-      '--option <name>=<value>',
-      'Pre-fill options for the creation process',
-      (opt, arr: string[]) => [...arr, opt],
-      [],
-    )
-    .option(
-      '--skip-install',
-      `Skips running 'yarn install' and 'yarn lint --fix'`,
-    )
-    .option('--scope <scope>', 'The scope to use for new packages')
-    .option(
-      '--npm-registry <URL>',
-      'The package registry to use for new packages',
-    )
-    .option(
-      '--baseVersion <version>',
-      'The version to use for any new packages (default: 0.1.0)',
-    )
-    .option(
-      '--license <license>',
-      'The license to use for any new packages (default: Apache-2.0)',
-    )
-    .option('--no-private', 'Do not mark new packages as private')
-    .action(lazy(() => import('./commands/new'), 'default'));
+export default createCliPlugin({
+  pluginId: 'new',
+  init: async reg => {
+    reg.addCommand({
+      path: ['new'],
+      description:
+        'Open up an interactive guide to creating new things in your app',
+      execute: async ({ args }) => {
+        const command = new Command();
+        const defaultCommand = command
+          .storeOptionsAsProperties(false)
+          .description(
+            'Open up an interactive guide to creating new things in your app',
+          )
+          .option(
+            '--select <name>',
+            'Select the thing you want to be creating upfront',
+          )
+          .option(
+            '--option <name>=<value>',
+            'Pre-fill options for the creation process',
+            (opt, arr: string[]) => [...arr, opt],
+            [],
+          )
+          .option(
+            '--skip-install',
+            `Skips running 'yarn install' and 'yarn lint --fix'`,
+          )
+          .option('--scope <scope>', 'The scope to use for new packages')
+          .option(
+            '--npm-registry <URL>',
+            'The package registry to use for new packages',
+          )
+          .option(
+            '--baseVersion <version>',
+            'The version to use for any new packages (default: 0.1.0)',
+          )
+          .option(
+            '--license <license>',
+            'The license to use for any new packages (default: Apache-2.0)',
+          )
+          .option('--no-private', 'Do not mark new packages as private')
+          .action(lazy(() => import('./commands/new'), 'default'));
 
-  program
-    .command('create')
-    .allowUnknownOption(true)
-    .action(removed("use 'backstage-cli new' instead"));
-  program
-    .command('create-plugin')
-    .allowUnknownOption(true)
-    .action(removed("use 'backstage-cli new' instead"));
-}
+        await defaultCommand.parseAsync(args, { from: 'user' });
+      },
+    });
+
+    reg.addCommand({
+      path: ['create'],
+      description: 'Create a new Backstage app',
+      deprecated: true,
+      execute: async () => {
+        removed("use 'backstage-cli new' instead")();
+      },
+    });
+    reg.addCommand({
+      path: ['create-plugin'],
+      description: 'Create a new Backstage plugin',
+      deprecated: true,
+      execute: async () => {
+        removed("use 'backstage-cli new' instead")();
+      },
+    });
+  },
+});
