+++ packages/cli/src/modules/test/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 The Backstage Authors
+ * Copyright 2024 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,39 +13,52 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
+import { createCliPlugin } from '../../wiring/factory';
 import { Command } from 'commander';
 import { lazy } from '../../lib/lazy';
 
-export function registerRepoCommands(command: Command) {
-  command
-    .command('test')
-    .allowUnknownOption(true) // Allows the command to run, but we still need to parse raw args
-    .option(
-      '--since <ref>',
-      'Only test packages that changed since the specified ref',
-    )
-    .option(
-      '--successCache',
-      'Enable success caching, which skips running tests for unchanged packages that were successful in the previous run',
-    )
-    .option(
-      '--successCacheDir <path>',
-      'Set the success cache location, (default: node_modules/.cache/backstage-cli)',
-    )
-    .option(
-      '--jest-help',
-      'Show help for Jest CLI options, which are passed through',
-    )
-    .description('Run tests, forwarding args to Jest, defaulting to watch mode')
-    .action(lazy(() => import('./commands/repo/test'), 'command'));
-}
+export default createCliPlugin({
+  pluginId: 'test',
+  init: async reg => {
+    reg.addCommand({
+      path: ['repo', 'test'],
+      description:
+        'Run tests, forwarding args to Jest, defaulting to watch mode',
+      execute: async ({ args }) => {
+        const command = new Command();
+        command.allowUnknownOption(true);
+        command.option(
+          '--since <ref>',
+          'Only test packages that changed since the specified ref',
+        );
+        command.option('--successCache', 'Enable success caching');
+        command.option(
+          '--successCacheDir <path>',
+          'Set the success cache location, (default: node_modules/.cache/backstage-cli)',
+        );
+        command.option(
+          '--jest-help',
+          'Show help for Jest CLI options, which are passed through',
+        );
+        command.action(lazy(() => import('./commands/repo/test'), 'command'));
+        await command.parseAsync(args, { from: 'user' });
+      },
+    });
 
-export function registerPackageCommands(command: Command) {
-  command
-    .command('test')
-    .allowUnknownOption(true) // Allows the command to run, but we still need to parse raw args
-    .helpOption(', --backstage-cli-help') // Let Jest handle help
-    .description('Run tests, forwarding args to Jest, defaulting to watch mode')
-    .action(lazy(() => import('./commands/package/test'), 'default'));
-}
+    reg.addCommand({
+      path: ['package', 'test'],
+      description:
+        'Run tests, forwarding args to Jest, defaulting to watch mode',
+      execute: async ({ args }) => {
+        const command = new Command();
+
+        command.allowUnknownOption(true);
+        command.helpOption(', --backstage-cli-help');
+        command.action(
+          lazy(() => import('./commands/package/test'), 'default'),
+        );
+        await command.parseAsync(args, { from: 'user' });
+      },
+    });
+  },
+});
